requirements.txt
aiogram
sqlalchemy
python-dotenv
asyncpg
asyncio